BOOKING-163: Removed unnecessary component layer
